requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
python-dotenv==1.0.0
openai==1.3.0
python-multipart==0.0.6
scikit-learn==1.3.2
numpy==1.26.2
pydantic==2.5.1
